chore: nimo requests
