--- packages/core/src/types/types.d.ts
@@ -1,2 +1,0 @@
-// TODO: find a way to put this in index.d.ts
-declare module "eigen";
