+++ packages/core/src/utils/Util.ts
@@ -318,7 +318,7 @@ export function mapMap(map: Map<any, any>, fn: any) {
  * @param argument Possible unsafe function call
  * @param message Error message
  */
-export const safe = <T extends {}>(
+export const safe = <T extends unknown>(
   argument: T | undefined | null,
   message: string
 ): T => {
