+++ packages/core/tsconfig.json
@@ -21,7 +21,6 @@
     "skipLibCheck": true,
     "downlevelIteration": true,
     "types": ["jest", "node", "tslib"],
-    "typeRoots": ["src/types", "node_modules/@types"],
     "esModuleInterop": true,
     "isolatedModules": true,
     "strict": true,
@@ -49,7 +48,7 @@
     "placeOpenBraceOnNewLineForFunctions": false,
     "placeOpenBraceOnNewLineForControlBlocks": false
   },
-  "include": ["src"],
+  "include": ["src", "index.d.ts"],
   "exclude": [
     "node_modules",
     "build",
